Move the HomeKit thermostat linked fan service into the climate base
