+++ tests/components/homekit/test_type_thermostats.py
@@ -44,6 +44,11 @@ from homeassistant.components.climate import (
     HVACMode,
 )
 from homeassistant.components.homekit.accessories import HomeDriver
+from homeassistant.components.homekit.climate_base import (
+    FAN_STATE_ACTIVE,
+    FAN_STATE_IDLE,
+    FAN_STATE_INACTIVE,
+)
 from homeassistant.components.homekit.const import (
     ATTR_VALUE,
     CHAR_CURRENT_FAN_STATE,
@@ -57,9 +62,6 @@ from homeassistant.components.homekit.const import (
     PROP_MIN_VALUE,
 )
 from homeassistant.components.homekit.type_thermostats import (
-    FAN_STATE_ACTIVE,
-    FAN_STATE_IDLE,
-    FAN_STATE_INACTIVE,
     HC_HEAT_COOL_AUTO,
     HC_HEAT_COOL_COOL,
     HC_HEAT_COOL_HEAT,
@@ -2336,13 +2338,13 @@ async def test_thermostat_with_fan_modes_with_auto(
     assert call_set_fan_mode[-1].data[ATTR_ENTITY_ID] == entity_id
     assert call_set_fan_mode[-1].data[ATTR_FAN_MODE] == FAN_LOW
 
-    char_active_iid = acc.char_active.to_HAP()[HAP_REPR_IID]
+    char_fan_active_iid = acc.char_fan_active.to_HAP()[HAP_REPR_IID]
     hk_driver.set_characteristics(
         {
             HAP_REPR_CHARS: [
                 {
                     HAP_REPR_AID: acc.aid,
-                    HAP_REPR_IID: char_active_iid,
+                    HAP_REPR_IID: char_fan_active_iid,
                     HAP_REPR_VALUE: 0,
                 }
             ]
@@ -2351,7 +2353,7 @@ async def test_thermostat_with_fan_modes_with_auto(
     )
 
     await hass.async_block_till_done()
-    assert acc.char_active.value == 1
+    assert acc.char_fan_active.value == 1
 
     char_target_fan_state_iid = acc.char_target_fan_state.to_HAP()[HAP_REPR_IID]
 
@@ -2434,7 +2436,7 @@ async def test_thermostat_with_fan_modes_with_off(
     assert CHAR_TARGET_FAN_STATE not in acc.fan_chars
     assert CHAR_SWING_MODE in acc.fan_chars
     assert CHAR_CURRENT_FAN_STATE in acc.fan_chars
-    assert acc.char_active.value == 1
+    assert acc.char_fan_active.value == 1
 
     hass.states.async_set(
         entity_id,
@@ -2460,16 +2462,16 @@ async def test_thermostat_with_fan_modes_with_off(
         },
     )
     await hass.async_block_till_done()
-    assert acc.char_active.value == 0
+    assert acc.char_fan_active.value == 0
 
     call_set_fan_mode = async_mock_service(hass, CLIMATE_DOMAIN, SERVICE_SET_FAN_MODE)
-    char_active_iid = acc.char_active.to_HAP()[HAP_REPR_IID]
+    char_fan_active_iid = acc.char_fan_active.to_HAP()[HAP_REPR_IID]
     hk_driver.set_characteristics(
         {
             HAP_REPR_CHARS: [
                 {
                     HAP_REPR_AID: acc.aid,
-                    HAP_REPR_IID: char_active_iid,
+                    HAP_REPR_IID: char_fan_active_iid,
                     HAP_REPR_VALUE: 1,
                 }
             ]
@@ -2487,7 +2489,7 @@ async def test_thermostat_with_fan_modes_with_off(
             HAP_REPR_CHARS: [
                 {
                     HAP_REPR_AID: acc.aid,
-                    HAP_REPR_IID: char_active_iid,
+                    HAP_REPR_IID: char_fan_active_iid,
                     HAP_REPR_VALUE: 0,
                 }
             ]
